requirements.txt
Flask==2.2.3
PyYAML==6.0
requests==2.31.0
python-dotenv==1.0.0